+++ utah-design-system-website/src/react/propTypesShapes/PaginationExamplePropsShape.js
@@ -1,0 +1,10 @@
+import PropTypes from 'prop-types';
+
+export default PropTypes.exact({
+  className: PropTypes.string,
+  id: PropTypes.string,
+  pageSize: PropTypes.string,
+  totalNumberItems: PropTypes.string,
+  value: PropTypes.string,
+  wrapInElement: PropTypes.oneOf(['div', 'nav']),
+});